refactor: organize utils
